Add tests for tosca.nodes.network.Network and tosca.nodes.network.Port

+++ testing/base.py
@@ -69,23 +69,25 @@ class BaseAnsibleProvider:
         assert hasattr(self, 'PROVIDER') is not None
         assert self.PROVIDER in PROVIDERS
 
-    def get_ansible_create_output(self, template, template_filename=None, extra=None):
+    def get_ansible_create_output(self, template, template_filename=None, extra=None, delete_template=True):
         if not template_filename:
             template_filename = self.testing_template_filename()
         self.write_template(self.prepare_yaml(template))
         r = common_translate(template_filename, False, self.PROVIDER, ANSIBLE, TEST, False, extra=extra)
         print(r)
-        self.delete_template(template_filename)
+        if delete_template:
+            self.delete_template(template_filename)
         playbook = self.parse_yaml(r)
         return playbook
 
-    def get_ansible_delete_output(self, template, template_filename=None, extra=None):
+    def get_ansible_delete_output(self, template, template_filename=None, extra=None, delete_template=True):
         if not template_filename:
             template_filename = self.testing_template_filename()
         self.write_template(self.prepare_yaml(template))
         r = common_translate(template_filename, False, self.PROVIDER, ANSIBLE, TEST, True, extra=extra)
         print(r)
-        self.delete_template(template_filename)
+        if delete_template:
+            self.delete_template(template_filename)
         playbook = self.parse_yaml(r)
         return playbook
